+++ src/rpc/server.ts
@@ -219,10 +219,11 @@ export class RpcServer {
    * Fetch the full account entry for a Stellar account.
    *
    * @param {string} address The public address of the account to load.
-   * @returns {Promise<xdr.AccountEntry>} A promise which resolves to the
-   *    {@link Account} object with a populated sequence number
+   * @returns {Promise<xdr.AccountEntry>} Resolves to the full on-chain account
+   *    entry
    *
-   * @see {@link https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLedgerEntries | getLedgerEntries docs}
+   * @see
+   * {@link https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLedgerEntries | getLedgerEntries docs}
    *
    * @example
    * const accountId = "GBZC6Y2Y7Q3ZQ2Y4QZJ2XZ3Z5YXZ6Z7Z2Y4QZJ2XZ3Z5YXZ6Z7Z2Y4";
@@ -245,6 +246,27 @@ export class RpcServer {
     }
   }
 
+  /**
+   * Fetch the full trustline entry for a Stellar account.
+   *
+   * @param {string} account  The public address of the account whose trustline it is
+   * @param {string} asset    The trustline's asset
+   * @returns {Promise<xdr.TrustLineEntry>} Resolves to the full on-chain trustline
+   *    entry
+   *
+   * @see
+   * {@link https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLedgerEntries | getLedgerEntries docs}
+   *
+   * @example
+   * const accountId = "GBZC6Y2Y7Q3ZQ2Y4QZJ2XZ3Z5YXZ6Z7Z2Y4QZJ2XZ3Z5YXZ6Z7Z2Y4";
+   * const asset = new Asset(
+   *  "USDC",
+   *  "GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA5"
+   * );
+   * server.getTrustline(accountId, asset).then((entry) => {
+   *   console.log(`{asset.toString()} balance for ${accountId}:", entry.balance().toString());
+   * });
+   */
   public async getTrustline(account: string, asset: Asset): Promise<xdr.TrustLineEntry> {
     const trustlineLedgerKey = xdr.LedgerKey.trustline(
       new xdr.LedgerKeyTrustLine({
@@ -258,6 +280,55 @@ export class RpcServer {
       return entry.val.trustLine();
     } catch (e) {
       throw new Error(`Trustline for ${asset.getCode()}:${asset.getIssuer()} not found for ${account}`);
+    }
+  }
+
+  /**
+   * Fetch the full trustline entry for a Stellar account.
+   *
+   * @param {string} id   The strkey (`B...`) or hex (`00000000abcde...`) of the
+   *    claimable balance to load
+   * @returns {Promise<xdr.ClaimableBalanceEntry>} Resolves to the full on-chain
+   *    claimable balance entry
+   *
+   * @see
+   * {@link https://developers.stellar.org/docs/data/rpc/api-reference/methods/getLedgerEntries | getLedgerEntries docs}
+   *
+   * @example
+   * const id = "00000000178826fbfe339e1f5c53417c6fedfe2c05e8bec14303143ec46b38981b09c3f9";
+   * server.getClaimableBalance(id).then((entry) => {
+   *   console.log(`Claimable balance {id.substr(0, 12)} has:`);
+   *   console.log(`  asset:  ${Asset.fromXDRObject(entry.asset()).toString()}`;
+   *   console.log(`  amount: ${entry.amount().toString()}`;
+   * });
+   */
+  public async getClaimableBalance(id: string): Promise<xdr.ClaimableBalanceEntry> {
+    let balanceId;
+    if (StrKey.isValidClaimableBalance(id)) {
+      let buffer = StrKey.decodeClaimableBalance(id);
+
+      // Pad the version byte to be a full int32 like in the XDR spec
+      let v = Buffer.concat([ Buffer.from('\x00\x00\x00'), buffer.subarray(0, 1) ])
+
+      // Slap on the rest of it and decode it
+      balanceId = xdr.ClaimableBalanceId.fromXDR(
+        Buffer.concat([ v, buffer.subarray(1) ])
+      );
+    } else if (id.match(/[a-f0-9]{72}/i)) {
+      balanceId = xdr.ClaimableBalanceId.fromXDR(id, "hex")
+    } else {
+      throw new TypeError(`expected 72-char hex ID or strkey, not ${id}`)
+    }
+
+    const trustlineLedgerKey = xdr.LedgerKey.claimableBalance(
+      new xdr.LedgerKeyClaimableBalance({ balanceId })
+    );
+
+    try {
+      const entry = await this.getLedgerEntry(trustlineLedgerKey);
+      return entry.val.claimableBalance();
+    } catch (e) {
+      throw new Error(`Claimable balance ${id} not found`);
     }
   }
 
@@ -1071,8 +1142,8 @@ export class RpcServer {
    *
    * This is a convenience wrapper around {@link Server.getLedgerEntries}.
    *
-   * @param {string}  contractId    the contract ID (string `C...`) whose
-   *    balance of `sac` you want to know
+   * @param {string}  address the contract (string `C...`) or account ID
+   *    (`G...`) whose balance of `sac` you want to know
    * @param {Asset}   sac     the built-in SAC token (e.g. `USDC:GABC...`) that
    *    you are querying from the given `contract`.
    * @param {string}  [networkPassphrase] optionally, the network passphrase to